ref changed config class name
