+++ include/Config.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-class ConfigTree;
 class ConfigParser;
+class ConfigTokenizer;
 
 #include <sys/stat.h>
 
@@ -9,14 +9,14 @@ class ConfigParser;
 
 #include "ConfigNode.hpp"
 #include "ConfigParser.hpp"
-#include "ConfigTree.hpp"
+#include "ConfigTokenizer.hpp"
 
 #define FILE_NAME "./config_file/default.conf"
 
 class Config {
        private:
-        ConfigParser parser_;
-        ConfigTree tree_;
+        ConfigTokenizer parser_;
+        ConfigParser tree_;
 
        public:
         explicit Config(const std::string& filename);
@@ -24,6 +24,6 @@ class Config {
 
         void printTree();
         void printTreeRecursion(ConfigNode* node, int depth = 0);
-        ConfigParser& getParser();
-        const ConfigTree& getTree() const;
+        ConfigTokenizer& getParser();
+        const ConfigParser& getTree() const;
 };
