+++ src/config/ConfigParser.cpp
@@ -1,49 +1,107 @@
 #include "ConfigParser.hpp"
 
-#include <sstream>
+#include "Validator.hpp"
 
-ConfigParser::ConfigParser(std::string& filename) { makeTokenList_(filename); }
-
-ConfigParser::~ConfigParser() {}
-
-const std::vector<Token>& ConfigParser::getTokens() const {
-        return (this->tokens_);
+ConfigParser::ConfigParser(const ConfigTokenizer& parser) : parser(parser) {
+        std::fill(keyFlag_, keyFlag_ + 16, 0);
+        makeConfTree_(parser);
 }
 
-void ConfigParser::makeTokenList_(std::string& filename) {
-        std::ifstream file(filename.c_str());
-        if (!file.is_open()) {
-                std::cerr << "Failed to open file: " << filename << std::endl;
-                std::exit(1);
+ConfigParser::~ConfigParser() {
+        if (this->root_) {
+                deleteTree_(this->root_);
+                this->root_ = NULL;
         }
+}
 
-        std::string line;
-        int lineCount = 0;
-        while (std::getline(file, line)) {
-                std::istringstream iss(line);
-                std::string oneLine;
+ConfigNode* ConfigParser::getRoot() const { return (this->root_); }
 
-                std::getline(iss, oneLine, '#');  // #以降を削除
-                oneLine.erase(
-                    0, oneLine.find_first_not_of(" \t"));  // 先頭の空白を削除
-                oneLine.erase(oneLine.find_last_not_of(" \t") +
-                              1);  // 末尾の空白を削除
-                lineCount++;
+void ConfigParser::makeConfTree_(const ConfigTokenizer& parser) {
+        this->layers_[0] = new ConfigNode(Token("root", 99));
+        this->root_ = this->layers_[0];
+        std::vector<Token> tokens = parser.getTokens();
 
-                char c = oneLine[oneLine.size() - 1];
-                if (!(oneLine.empty()|| c == '{' || c == '}' || c == ';')) {
-                        std::cerr
-                            << "Syntax error at the end of the line: line "
-                            << lineCount << std::endl;
-                        std::exit(1);
-                }
+        for (size_t i = 0; i < tokens.size(); ++i) {
+                TokenType type = tokens[i].getType();
+                const std::string token = tokens[i].getText();
+                int depth = this->keyFlag_[BRACE];
+                int lineNumber = tokens[i].getLineNumber();
+                Validator::checkSyntaxErr(tokens[i], depth);
 
-                std::istringstream tokenStream(oneLine);
-                std::string token;
-                while (tokenStream >> token) {  // 空白区切りでtokenをset
-                        Token newToken(token, lineCount);
-                        this->tokens_.push_back(newToken);
-                }
+                if (type == BRACE)
+                        updateDepth_(token, lineNumber);
+                else if (type >= SERVER && type <= RETURN)
+                        addChild_(tokens[i], layers_[depth + 1],
+                                  layers_[depth]);
+                else if (i > 0 && tokens[i - 1].getType() == ERR_PAGE)
+                        addChild_(tokens[i], layers_[depth + 2],
+                                  layers_[depth + 1]);
+                else if (this->keyFlag_[ERR_PAGE] == 1)
+                        setValue_(tokens[i], layers_[depth + 2]);
+                else
+                        setValue_(tokens[i], layers_[depth + 1]);
         }
-        file.close();
+}
+
+void ConfigParser::updateDepth_(const std::string& token,
+                                const int lineNumber) {
+        if (token == "{") {
+                this->keyFlag_[BRACE]++;
+        } else if (token == "}") {
+                this->keyFlag_[BRACE]--;
+                if (this->keyFlag_[BRACE] < 0)
+                        errExit_(token, ":Config brace close error: line ",
+                                 lineNumber);
+                if (this->keyFlag_[BRACE] == 0 && keyFlag_[LOCATION] == 0)
+                        errExit_(token, ":Config location error: line ",
+                                 lineNumber);
+                if (this->keyFlag_[BRACE] == 0) resetKeyFlag_(LOCATION);
+                if (this->keyFlag_[BRACE] == 0) resetKeyFlag_(SERVER);
+        }
+}
+
+void ConfigParser::resetKeyFlag_(const int keyType) {
+        this->keyFlag_[keyType] = 0;
+}
+
+void ConfigParser::addChild_(const Token& token, ConfigNode*& current,
+                             ConfigNode* parent) {
+        current = new ConfigNode(token);
+        parent->getChildren().push_back(current);
+        int keyType = token.getType();
+        this->keyFlag_[keyType]++;
+}
+
+void ConfigParser::setValue_(const Token& token, ConfigNode* node) {
+        std::string text = token.getText();
+        Validator::checkSyntaxErr(token, this->keyFlag_[BRACE]);
+
+        if (text.size() == 1 && text[0] == ';')
+                errExit_(text, ": Can't find value: line ",
+                         token.getLineNumber());
+        if (text[text.size() - 1] == ';') {
+                text = text.substr(0, text.size() - 1);
+                if (this->keyFlag_[ERR_PAGE] == 1) resetKeyFlag_(ERR_PAGE);
+        }
+        node->getValues().push_back(text);
+}
+
+void ConfigParser::errExit_(const std::string& str1, const std::string& str2,
+                            const int number) {
+        std::cerr << str1 << str2 << number << std::endl;
+        if (this->root_) {
+                deleteTree_(this->root_);
+                this->root_ = NULL;
+        }
+        std::exit(1);
+}
+
+void ConfigParser::deleteTree_(ConfigNode* node) {
+        if (!node) return;
+        for (std::vector<ConfigNode*>::iterator it =
+                 node->getChildren().begin();
+             it != node->getChildren().end(); ++it)
+                deleteTree_(*it);
+        delete (node);
+        node = NULL;
 }
