+++ src/config/ConfigTokenizer.cpp
@@ -1,0 +1,51 @@
+#include "ConfigTokenizer.hpp"
+
+#include <sstream>
+
+ConfigTokenizer::ConfigTokenizer(std::string& filename) {
+        makeTokenList_(filename);
+}
+
+ConfigTokenizer::~ConfigTokenizer() {}
+
+const std::vector<Token>& ConfigTokenizer::getTokens() const {
+        return (this->tokens_);
+}
+
+void ConfigTokenizer::makeTokenList_(std::string& filename) {
+        std::ifstream file(filename.c_str());
+        if (!file.is_open()) {
+                std::cerr << "Failed to open file: " << filename << std::endl;
+                std::exit(1);
+        }
+
+        std::string line;
+        int lineCount = 0;
+        while (std::getline(file, line)) {
+                std::istringstream iss(line);
+                std::string oneLine;
+
+                std::getline(iss, oneLine, '#');  // #以降を削除
+                oneLine.erase(
+                    0, oneLine.find_first_not_of(" \t"));  // 先頭の空白を削除
+                oneLine.erase(oneLine.find_last_not_of(" \t") +
+                              1);  // 末尾の空白を削除
+                lineCount++;
+
+                char c = oneLine[oneLine.size() - 1];
+                if (!(oneLine.empty() || c == '{' || c == '}' || c == ';')) {
+                        std::cerr
+                            << "Syntax error at the end of the line: line "
+                            << lineCount << std::endl;
+                        std::exit(1);
+                }
+
+                std::istringstream tokenStream(oneLine);
+                std::string token;
+                while (tokenStream >> token) {  // 空白区切りでtokenをset
+                        Token newToken(token, lineCount);
+                        this->tokens_.push_back(newToken);
+                }
+        }
+        file.close();
+}
